added tester

+++ headers/Iterator_traits.hpp
@@ -32,8 +32,8 @@ struct	ft::iterator_traits<ft::base_iterator<ft::bidirectional_iterator_tag, T> 
 template <class T>
 struct	ft::iterator_traits<T*>
 {
-	typedef	ptrdiff_t					difference_type;
-	// typedef	long						difference_type;
+	// typedef	ptrdiff_t					difference_type;
+	typedef	long						difference_type;
 	typedef	T							value_type;
 	typedef	T*							pointer;
 	typedef	T&							reference;
@@ -43,8 +43,8 @@ struct	ft::iterator_traits<T*>
 template <class T>
 struct	ft::iterator_traits<const T*>
 {
-	typedef	ptrdiff_t					difference_type;
-	// typedef	long						difference_type;
+	// typedef	ptrdiff_t					difference_type;
+	typedef	long						difference_type;
 	typedef	T							value_type;
 	typedef	const T*					pointer;
 	typedef const T&					reference;
